refactor(pii): reorganize blurred-pii lib
