+++ lib/blurred-pii/context/BlurredPIIContext.tsx
@@ -1,0 +1,4 @@
+import * as React from "react"
+
+const defaultShouldBlur = false
+export const BlurredPIIContext = React.createContext(defaultShouldBlur)
