+++ lib/blurred-pii/hook.ts
@@ -1,0 +1,12 @@
+import * as React from "react"
+import { withBlurredPII, BlurredPIIOptions } from "./helpers"
+
+interface UseBlurredPII {
+  withBlurredPII: (options?: BlurredPIIOptions) => React.CSSProperties
+}
+
+export const useBlurredPII = (): UseBlurredPII => {
+  return {
+    withBlurredPII,
+  }
+}
